requirements.txt
Django==1.9.5
django-uuslug==1.1.7
enum34==1.1.4
psycopg2==2.6.1
python-slugify==1.2.0
Unidecode==0.4.19
wheel==0.24.0
